refactor(tests): address PR review — extract multi-dim tests, fix assert scope
- Remove `assert D % 2 == 0` from generic `triton_flash_attention` wrapper
  (only TQ4 kernels require even HEAD_DIM for nibble packing; the vanilla
  FA kernel supports any positive HEAD_DIM via padding+masking)
- Extract `TestMultiDimTritonEquivalence` from vLLM-gated
  `test_vllm_triton_equivalence.py` to new `test_triton_multi_dim.py`
  (multi-dim tests don't use vLLM; module-level importorskip was
  unnecessarily hiding 9 tests from non-vLLM environments)
